Finished iclient lock functionality
Note that lease RPC failures (+ demotions + revocations) not yet handled.

Also fixed bad OpenHandle bug... If two FHs were created for a since
Inode, a subsequent Lookup-by-InodeNumber would only find the 2nd. The
fix is simply to only Lookup-by-FissionFH... and not deal with the
unnecessary complexity of tracking all FHs for a given InodeNumber.

Instead, we now simply keep a count of # of FissionFH's for each Inode.